WoT S20 (part 1): move scroll + viewport files into slice folders
git mv into vertical-slice folders + import-path fixes only (no logic change):
- scroll.ts -> scroll/scroll.ts
- viewport.ts, workspaceSize.ts, calculatorFactory.ts, renderedRange.ts -> viewport/

Intra-slice imports stay same-folder; cross-slice + importer paths bumped. Git
tracks all five as renames; the diff is renames + import lines only.

test:walkontable 745/0/2, test:types + eslint green.
