+++ handsontable/src/3rdparty/walkontable/src/viewport/workspaceSize.ts
